[skip] Removed validation of namespaces (1h)
Before, namespace validation required traversing up the tree to find the ancestor with the namespace definition. When the ancestor is the document root and every element at every level is namespaced, this becomes problematic very quickly. Since this gem is not intended as a general-purpose XML builder, but as something to be used by other openoffice format-generating gems, we can assume that namespaces will be correctly defined in the proper places in those gems, making the need for verification superfluous, meaning we can remove it altogether and speed things up considerably.