Allow relationships to be built outside of rels
This is because in some cases it is nice to not build a new rels each
time you push a rel. This means that the rels can have the same ref ids
no matter what rel file they are in. This will also reduce memory
footprint and time.

Amos King @adkron <amos@binarynoggin.com>