[improvement] Add validation param to attribute (5m)
Adding attributes to an anonymous subclass of `OpenXml::Element` to facilitate properties attributes introduces an interesting wrinkle: `expects` no longer has access to custom validation methods! Since the scope of the `attribute` call is the anonymous class, any validation methods on the actual named element with properties is out of scope! Adding `validation`, which accepts anything that can be `call`ed, works around this by allowing us to pass in a closure-of-sorts to do custom validation on an attribute.